ENTIDAD CONFIGURATION CREADA
Creada la entidad configuration la cual sustituye la entidad spamWord y
Umbral para así poder tener todos los parámetros de configuración en una
misma clase.
