--- src/main/java/acme/entities/spamWord/SpamWord.java
@@ -1,20 +1,0 @@
-package acme.entities.spamWord;
-
-import javax.persistence.Entity;
-import javax.validation.constraints.NotBlank;
-
-import acme.framework.entities.DomainEntity;
-import lombok.Getter;
-import lombok.Setter;
-
-@Entity
-@Getter
-@Setter
-public class SpamWord extends DomainEntity{
-	
-	protected static final long serialVersionUID = 1L;
-	
-	@NotBlank
-	protected String word;
-
-}
