+++ src/main/java/acme/features/manager/task/ManagerTaskCreateService.java
@@ -7,7 +7,7 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import acme.entities.tasks.Task;
-import acme.features.spamWord.SpamWordFilterService;
+import acme.features.configuration.ConfigurationService;
 import acme.framework.components.Errors;
 import acme.framework.components.Model;
 import acme.framework.components.Request;
@@ -22,7 +22,7 @@ public class ManagerTaskCreateService implements AbstractCreateService<Manager, 
 	protected ManagerTaskRepository repository;
 	
 	@Autowired
-	protected SpamWordFilterService spamService;
+	protected ConfigurationService confService;
 	
 	@Override
 	public boolean authorise(final Request<Task> request) {
@@ -94,7 +94,7 @@ public class ManagerTaskCreateService implements AbstractCreateService<Manager, 
 		}
 		
 		if(!errors.hasErrors("description")) {
-			final boolean umbralSuperado = this.spamService.spamFilter(entity.getDescription(), 10);
+			final boolean umbralSuperado = this.confService.spamFilter(entity.getDescription(), 10);
 			errors.state(request, !umbralSuperado, "description", "manager.task.error.umbral-superado");
 		}
