--- src/main/java/acme/features/spamWord/SpamWordRepository.java
@@ -1,14 +1,0 @@
-package acme.features.spamWord;
-
-import java.util.List;
-
-import org.springframework.data.jpa.repository.Query;
-
-import acme.entities.spamWord.SpamWord;
-import acme.framework.repositories.AbstractRepository;
-
-public interface SpamWordRepository extends AbstractRepository{
-
-	@Query("select s from SpamWord s")
-	List<SpamWord> findMany();
-}
